updating imports w.r.t. to moving  into  (RSE-Sheffield/deepquest#43).
