+++ deepquest/predict.py
@@ -11,7 +11,7 @@ from keras_wrapper.dataset import loadDataset
 from . import qe_models as modFactory
 from .utils import evaluation
 from .utils.callbacks import *
-from .utils.prepare_data import build_dataset, update_dataset_from_file, keep_n_captions
+from .data_engine.prepare_data import build_dataset, update_dataset_from_file, keep_n_captions
 
 logging.basicConfig(level=logging.DEBUG,
                     format='[%(asctime)s] %(message)s', datefmt='%d/%m/%Y %H:%M:%S')
